Make a distinct analytics folder
