+++ ydb/tests/fq/generic/analytics/clickhouse/init.sql
