+++ ydb/tests/fq/generic/analytics/docker-compose.yml
