+++ ydb/tests/fq/generic/analytics/greenplum/init_db.sh
