+++ ydb/tests/fq/generic/analytics/postgresql/init.sql
