+++ ydb/tests/fq/generic/analytics/test_postgresql.py
@@ -6,29 +6,29 @@ import ydb.public.api.protos.ydb_value_pb2 as ydb
 from ydb.tests.tools.fq_runner.kikimr_utils import yq_v2
 
 from ydb.tests.tools.fq_runner.fq_client import FederatedQueryClient
-from ydb.tests.fq.generic.utils.settings import Settings
+from ydb.tests.fq.generic.analytics.utils.settings import Settings
 
 
-class TestGreenplum:
+class TestPostgreSQL:
     @yq_v2
     @pytest.mark.parametrize("fq_client", [{"folder_id": "my_folder"}], indirect=True)
     def test_simple(self, fq_client: FederatedQueryClient, settings: Settings):
-        table_name = "simple_table"
-        conn_name = f"conn_{table_name}"
-        query_name = f"query_{table_name}"
+        table_name = 'simple_table'
+        conn_name = f'conn_{table_name}'
+        query_name = f'query_{table_name}'
 
-        fq_client.create_greenplum_connection(
+        fq_client.create_postgresql_connection(
             name=conn_name,
-            database_name=settings.greenplum.dbname,
-            database_id="greenplum_cluster_id",
-            login=settings.greenplum.username,
-            password=settings.greenplum.password,
+            database_name=settings.postgresql.dbname,
+            database_id='postgresql_cluster_id',
+            login=settings.postgresql.username,
+            password=settings.postgresql.password,
         )
 
-        sql = Rf"""
+        sql = fR'''
             SELECT *
-            FROM {conn_name}.{table_name} ORDER BY number;
-            """
+            FROM {conn_name}.{table_name};
+            '''
 
         query_id = fq_client.create_query(query_name, sql, type=fq.QueryContent.QueryType.ANALYTICS).result.query_id
         fq_client.wait_query_status(query_id, fq.QueryMeta.COMPLETED)
@@ -42,6 +42,6 @@ class TestGreenplum:
             optional_type=ydb.OptionalType(item=ydb.Type(type_id=ydb.Type.INT32))
         )
         assert len(result_set.rows) == 3
-        assert result_set.rows[0].items[0].int32_value == 3
-        assert result_set.rows[1].items[0].int32_value == 14
-        assert result_set.rows[2].items[0].int32_value == 15
+        assert result_set.rows[0].items[0].int32_value == 1
+        assert result_set.rows[1].items[0].int32_value == 2
+        assert result_set.rows[2].items[0].int32_value == 3
