+++ ydb/tests/fq/generic/analytics/utils/endpoint_determiner.py
