+++ ydb/tests/fq/generic/analytics/ydb/01_basic.sh
